+++ scripts.js
@@ -118,7 +118,7 @@ var updateModule = function updateModule(module_name, callback) {
    }
 };
 
-var channel, contributors, mutes, mbans, smutes, detained, hmutes, mafiaSuperAdmins, hangmanAdmins, hangmanSuperAdmins, staffchannel, channelbot, normalbot, bot, mafiabot, kickbot, capsbot, checkbot, coinbot, countbot, tourneybot, battlebot, commandbot, querybot, rankingbot, hangbot, bfbot, scriptChecks, lastMemUpdate, bannedUrls, mafiachan, mafiarev, sachannel, tourchannel, dwpokemons, hapokemons, lcpokemons, bannedGSCSleep, bannedGSCTrap, breedingpokemons, rangebans, proxy_ips, mafiaAdmins, rules, nameBans, chanNameBans, isSuperAdmin, cmp, key, battlesStopped, lineCount, pokeNatures, pokeAbilities, maxPlayersOnline, pastebin_api_key, pastebin_user_key, getSeconds, getTimeString, sendChanMessage, sendChanAll, sendMainTour, VarsCreated, authChangingTeam, usingBannedWords, repeatingOneself, capsName, CAPSLOCKDAYALLOW, nameWarns, poScript, revchan, triviachan, watchchannel, lcmoves, hangmanchan, ipbans, battlesFought, lastCleared, blackjackchan, heightList, weightList, powerList, accList, ppList, categoryList, moveEffList, moveFlagList, abilityList, itemList, berryList, flingPowerList, berryPowerList, berryTypeList, allMovesList, allGenMovesList, namesToWatch, allowedRangeNames, reverseTohjo;
+var channel, contributors, mutes, mbans, smutes, detained, hmutes, mafiaSuperAdmins, hangmanAdmins, hangmanSuperAdmins, staffchannel, channelbot, normalbot, bot, mafiabot, kickbot, capsbot, checkbot, coinbot, countbot, tourneybot, battlebot, commandbot, querybot, rankingbot, hangbot, bfbot, scriptChecks, lastMemUpdate, bannedUrls, mafiachan, mafiarev, sachannel, tourchannel, dwpokemons, hapokemons, lcpokemons, bannedGSCSleep, bannedGSCTrap, breedingpokemons, rangebans, proxy_ips, mafiaAdmins, rules, authStats, nameBans, chanNameBans, isSuperAdmin, cmp, key, battlesStopped, lineCount, pokeNatures, pokeAbilities, maxPlayersOnline, pastebin_api_key, pastebin_user_key, getSeconds, getTimeString, sendChanMessage, sendChanAll, sendMainTour, VarsCreated, authChangingTeam, usingBannedWords, repeatingOneself, capsName, CAPSLOCKDAYALLOW, nameWarns, poScript, revchan, triviachan, watchchannel, lcmoves, hangmanchan, ipbans, battlesFought, lastCleared, blackjackchan, namesToWatch, allowedRangeNames, reverseTohjo;
 
 var pokeDir = "db/pokes/";
 var moveDir = "db/moves/6G/";
@@ -127,7 +127,7 @@ var itemDir = "db/items/";
 sys.makeDir("scripts");
 /* we need to make sure the scripts exist */
 var commandfiles = ['commands.js', 'channelcommands.js','ownercommands.js', 'modcommands.js', 'usercommands.js', "admincommands.js"];
-var deps = ['crc32.js', 'utilities.js', 'bot.js', 'memoryhash.js', 'tierchecks.js', "globalfunctions.js", "userfunctions.js", "channelfunctions.js", "channelmanager.js"].concat(commandfiles).concat(Config.Plugins);
+var deps = ['crc32.js', 'utilities.js', 'bot.js', 'memoryhash.js', 'tierchecks.js', "globalfunctions.js", "userfunctions.js", "channelfunctions.js", "channelmanager.js", "pokedex.js"].concat(commandfiles).concat(Config.Plugins);
 var missing = 0;
 for (var i = 0; i < deps.length; ++i) {
     if (!sys.getFileContent("scripts/"+deps[i])) {
@@ -156,6 +156,8 @@ var POUser = require('userfunctions.js').POUser;
 var POGlobal = require('globalfunctions.js').POGlobal;
 delete require.cache['tierchecks.js'];
 var tier_checker = require('tierchecks.js');
+delete require.cache['pokedex.js'];
+var pokedex = require('pokedex.js');
 
 /* stolen from here: http://stackoverflow.com/questions/610406/javascript-equivalent-to-printf-string-format */
 String.prototype.format = function() {
@@ -165,6 +167,15 @@ String.prototype.format = function() {
         formatted = formatted.replace(regexp, arguments[i]);
     }
     return formatted;
+};
+
+String.prototype.toCorrectCase = function() {
+    if (isNaN(this) && sys.id(this) !== undefined) {
+        return sys.name(sys.id(this));
+    }
+    else {
+        return this;
+    }
 };
 
 var utilities = require('utilities.js');
@@ -222,274 +233,6 @@ function sendChanHtmlAll(message, chan_id) {
     {
         sys.sendHtmlAll(message, chan_id);
     }
-}
-
-String.prototype.toCorrectCase = function() {
-    if (isNaN(this) && sys.id(this) !== undefined) {
-        return sys.name(sys.id(this));
-    }
-    else {
-        return this;
-    }
-};
-function dwCheck(pokemon){
-    if (sys.pokeAbility(pokemon,2,5) === 0 && sys.pokeAbility(pokemon,1,5) === 0){
-        return false;
-    }
-    return true;
-}
-
-function calcStat (base, IV, EV, level, nature) {
-    var stat = Math.floor(Math.floor((IV + (2 * base) + Math.floor(EV / 4)) * level / 100 + 5) * nature);
-    return stat;
-}
-
-function calcHP (base, IV, EV, level) {
-    if (base === 1) {
-        return 1;
-    }
-    var HP = Math.floor((IV + (2 * base) + Math.floor(EV / 4) + 100) * level / 100 + 10);
-    return HP;
-}
-
-function getDBIndex (pokeId) {
-    var id = pokeId % 65536;
-    var forme = (pokeId - id) / 65536;
-    return id + ":" + forme;
-}
-
-function getWeight (pokeId) {
-    if (weightList === undefined) {
-        weightList = {};
-        var data = sys.getFileContent(pokeDir + 'weight.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var id = data[i].substr(0, index);
-            var weight = data[i].substr(index + 1);
-            weightList[id] = weight;
-        }
-    }
-    var key = getDBIndex(pokeId);
-    if (weightList[key] !== undefined) {
-        return weightList[key];
-    }
-    var index = key.indexOf(":") + 1;
-    var base = key.substr(0, index);
-    return weightList[base + "0"];
-}
-
-function getHeight (pokeId) {
-    if (heightList === undefined) {
-        heightList = {};
-        var data = sys.getFileContent(pokeDir + 'height.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var id = data[i].substr(0, index);
-            var height = data[i].substr(index + 1);
-            heightList[id] = height;
-        }
-    }
-    var key = getDBIndex(pokeId);
-    if (heightList[key] !== undefined) {
-        return heightList[key];
-    }
-    var index = key.indexOf(":") + 1;
-    var base = key.substr(0, index);
-    return heightList[base + "0"];
-}
-
-function weightPower (weight) {
-    var power = 0;
-    if (weight < 10) power = 20;
-    if (weight >= 10 && weight < 25) power = 40;
-    if (weight >= 25 && weight < 50) power = 60;
-    if (weight >= 50 && weight < 100) power = 80;
-    if (weight >= 100 && weight < 200) power = 100;
-    if (weight >= 200) power = 120;
-    return power;
-}
-
-function getMoveBP (moveId) {
-    if (powerList === undefined) {
-        powerList = {};
-        var data = sys.getFileContent(moveDir + 'power.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var power = data[i].substr(index + 1);
-            powerList[key] = power;
-        }
-    }
-    if (powerList[moveId] === undefined || powerList[moveId] === "1") {
-        return "---";
-    }
-    return powerList[moveId];
-}
-
-function getMoveCategory (moveId) {
-    if (categoryList === undefined) {
-        categoryList = {};
-        var data = sys.getFileContent(moveDir + 'damage_class.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var category = data[i].substr(index + 1);
-            categoryList[key] = category;
-        }
-    }
-    if (categoryList[moveId] === "1") {
-        return "Physical";
-    }
-    if (categoryList[moveId] === "2") {
-        return "Special";
-    }
-    return "Other";
-}
-
-function getMoveAccuracy (moveId) {
-    if (accList === undefined) {
-        accList = {};
-        var data = sys.getFileContent(moveDir + 'accuracy.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var accuracy = data[i].substr(index + 1);
-            accList[key] = accuracy;
-        }
-    }
-    if (accList[moveId] === "101") {
-        return "---";
-    }
-    return accList[moveId];
-}
-
-function getMovePP (moveId) {
-    if (ppList === undefined) {
-        ppList = {};
-        var data = sys.getFileContent(moveDir + 'pp.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var pp = data[i].substr(index + 1);
-            ppList[key] = pp;
-        }
-    }
-    return ppList[moveId];
-}
-
-function getMoveEffect (moveId) {
-    if (moveEffList === undefined) {
-        moveEffList = {};
-        var data = sys.getFileContent(moveDir + 'effect.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var effect = data[i].substr(index + 1);
-            moveEffList[key] = effect;
-        }
-    }
-    if (moveEffList[moveId] === undefined) {
-        return "Deals normal damage.";
-    }
-    return moveEffList[moveId].replace(/[\[\]{}]/g, "");
-}
-
-function getMoveContact (moveId) {
-    if (moveFlagList === undefined) {
-        moveFlagList = {};
-        var data = sys.getFileContent(moveDir + 'flags.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var flags = data[i].substr(index + 1);
-            moveFlagList[key] = flags;
-        }
-    }
-    return moveFlagList[moveId] % 2 === 1;
-}
-
-function getAbility (abilityId) {
-    if (abilityList === undefined) {
-        abilityList = {};
-        var data = sys.getFileContent(abilityDir + 'ability_battledesc.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var ability = data[i].substr(index + 1);
-            abilityList[key] = ability;
-        }
-    }
-    return abilityList[abilityId];
-}
-
-function getItem (itemId) {
-    if (itemList === undefined) {
-        itemList = {};
-        var data = sys.getFileContent(itemDir + 'items_description.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var item = data[i].substr(index + 1);
-            itemList[key] = item;
-        }
-    }
-    return itemList[itemId];
-}
-
-function getBerry (berryId) {
-    if (berryList === undefined) {
-        berryList = {};
-        var data = sys.getFileContent(itemDir + 'berries_description.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var berry = data[i].substr(index + 1);
-            berryList[key] = berry;
-        }
-    }
-    return berryList[berryId];
-}
-
-function getFlingPower (itemId) {
-    if (flingPowerList === undefined) {
-        flingPowerList = {};
-        var data = sys.getFileContent(itemDir + 'items_pow.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var power = data[i].substr(index + 1);
-            flingPowerList[key] = power;
-        }
-    }
-    return flingPowerList[itemId];
-}
-
-function getBerryPower (berryId) {
-    if (berryPowerList === undefined) {
-        berryPowerList = {};
-        var data = sys.getFileContent(itemDir + 'berry_pow.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var power = data[i].substr(index + 1);
-            berryPowerList[key] = power;
-        }
-    }
-    return +berryPowerList[berryId] + 20;
-}
-
-function getBerryType (berryId) {
-    if (berryTypeList === undefined) {
-        berryTypeList = {};
-        var data = sys.getFileContent(itemDir + 'berry_type.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var type = data[i].substr(index + 1);
-            berryTypeList[key] = sys.type(type);
-        }
-    }
-    return berryTypeList[berryId];
 }
 
 function updateNotice() {
@@ -2489,55 +2232,5 @@ hasAuthElements: function (array) {
         }
     }
     return false;
-},
-
-hasDreamWorldAbility: function (pokemon, ability) {
-    return sys.pokeAbility(pokemon, 2) === ability && sys.pokeAbility(pokemon, 0) !== sys.pokeAbility(pokemon, 2) && sys.pokeAbility(pokemon, 1) !== sys.pokeAbility(pokemon, 2);
-},
-
-getAllMoves: function (pokeId) {
-    if (allMovesList === undefined) {
-        allMovesList = {};
-        var data = sys.getFileContent('db/pokes/6G/all_moves.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var allMoves = data[i].substr(index + 1).split(" ");
-            allMovesList[key] = allMoves;
-        }
-    }
-    return allMovesList[getDBIndex(pokeId)];
-},
-
-getAllGenMoves: function (pokeId) {
-    if (!allGenMovesList) {
-        allGenMovesList = {};
-        var data = sys.getFileContent(Config.dataDir+'all_gen_moves.txt').split('\n');
-        for (var i = 0; i < data.length; i++) {
-            var index = data[i].indexOf(" ");
-            var key = data[i].substr(0, index);
-            var allGenMoves = data[i].substr(index + 1).split(" ");
-            allGenMovesList[key] = allGenMoves;
-        }
-    }
-    return allGenMovesList[getDBIndex(pokeId)];
-},
-
-natures: [["Hardy", "Lonely", "Adamant", "Naughty", "Brave"],
-	["Bold", "Docile", "Impish", "Lax", "Relaxed"],
-	["Modest", "Mild", "Bashful", "Rash", "Quiet"],
-	["Calm", "Gentle", "Careful", "Quirky", "Sassy"],
-	["Timid", "Hasty", "Jolly", "Naive", "Serious"]],
-
-getNatureEffect: function (nature) {
-    nature = nature.toLowerCase();
-    for (var x = 0; x < 5; x++) {
-        for (var y = 0; y < 5; y++) {
-            if (this.natures[x][y].toLowerCase() === nature) {
-                return [x, y];
-            }
-        }
-    }
-    return;
 }
 });
