+++ scripts/tierchecks.js
@@ -145,7 +145,7 @@ tier_checker.add_new_check(EXCLUDING, Config.DreamWorldTiers, function dwAbility
     if (sys.gen(src, team) === 6) {
         for (var i = 0; i < 6; i++) {
             var x = sys.teamPoke(src, team, i);
-            if (x !== 0 && script.hasDreamWorldAbility(x, sys.teamPokeAbility(src, team, i))) {
+            if (x !== 0 && pokedex.hasDreamWorldAbility(x, sys.teamPokeAbility(src, team, i))) {
                 if (!(x in script.hapokemons) || ((tier === "Pre-PokeBank OU" || tier === "Random Battle") && pokebank.indexOf(sys.pokemon(x)) !== -1)) {
                     ret.push("" + sys.pokemon(x) + " is not allowed with Hidden Ability " + sys.ability(sys.teamPokeAbility(src, team, i)) + " in " + tier + " tier. Change it in the teambuilder.");
                 }
@@ -410,7 +410,7 @@ tier_checker.add_new_check(INCLUDING, ["Pre-PokeBank OU"], function pokeBankChec
     for (var slot = 0; slot < 6; slot++) {
         var poke = sys.teamPoke(src, team, slot);
         if (poke) { 
-            var moves = script.getAllGenMoves(poke);
+            var moves = pokedex.getAllGenMoves(poke);
             if (moves) {
                 for (var move = 0; move < 4; move++) {
                     if (sys.teamPokeMove(src, team, slot, move)) {
